remove globals
