+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sunker-plugin-e2e",
-  "version": "1.0.35",
+  "version": "1.0.53",
   "description": "",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
