+++ src/e2e-selectors/index.ts
@@ -1,2 +1,3 @@
 export { resolveSelectors } from './resolver';
 export * from './types';
+export * from './versioned';
