+++ src/e2e-selectors/versioned/index.ts
@@ -1,2 +1,3 @@
+export { versionedAPIs } from './apis';
 export { versionedComponents } from './components';
 export { versionedPages } from './pages';
