+++ src/fixtures/selectors.ts
@@ -1,8 +1,7 @@
 import { TestFixture } from '@playwright/test';
-import { versionedAPIs } from 'src/e2e-selectors/versioned/apis';
 import { PluginFixture, PluginOptions } from '../api';
 import { E2ESelectors, resolveSelectors } from '../e2e-selectors';
-import { versionedComponents, versionedPages } from '../e2e-selectors/versioned';
+import { versionedAPIs, versionedComponents, versionedPages } from '../e2e-selectors/versioned';
 import { PlaywrightCombinedArgs } from './types';
 
 type SelectorFixture = TestFixture<E2ESelectors, PluginFixture & PluginOptions & PlaywrightCombinedArgs>;
